Commit all
