+++ Main/Main.java
@@ -1,7 +1,7 @@
-package main;
+package Main;
 
 public class Main {
     public static void main(String[] args) {
-        
+        System.out.println("test");
     }
 }
