+++ View/Create.java
@@ -148,7 +148,6 @@ public class Create {
         
             controller.CreateEvent.EventCreate(name, eventStart, eventEnd, desc, kategori, capacityText, priceText);
             frame.dispose();
-            new MainMenuSeller();
         });
