--- View/MainMenuCustomer.java
@@ -1,5 +1,0 @@
-package view;
-
-public class MainMenuCustomer {
-    
-}
